Moved email templates enum into config package, as per code-review comments.
